requirements.txt
discord.py[voice] >= 0.16.6
Django >= 1.11
mysqlclient >= 1.3.9
youtube_dl >= 2017.7.23
